+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/components/dashboard/main.js
@@ -21,16 +21,10 @@ define(function() {
         view: true,
 
         layout: {
-            navigation: {
-                collapsed: false
-            },
             content: {
                 width: 'max',
-                topSpace: false,
-                leftSpace: false,
-                rightSpace: false
-            },
-            sidebar: false
+                topSpace: false
+            }
         },
 
         header: {