+++ src/Sulu/Bundle/CategoryBundle/Resources/public/js/components/categories/list/main.js
@@ -23,9 +23,7 @@ define(function () {
 
         layout: {
             content: {
-                width: 'max',
-                leftSpace: false,
-                rightSpace: false
+                width: 'max'
             }
         },
 
@@ -50,9 +48,9 @@ define(function () {
         },
 
         bindCustomEvents: function() {
+            this.sandbox.on('husky.datagrid.item.click', this.saveLastClickedCategory.bind(this))
             this.sandbox.on('sulu.list-toolbar.add', this.addNewCategory.bind(this));
             this.sandbox.on('sulu.list-toolbar.delete', this.deleteSelected.bind(this));
-            this.sandbox.on('husky.datagrid.item.click', this.saveLastClickedCategory.bind(this));
         },
 
         /**
@@ -76,20 +74,15 @@ define(function () {
                     resultKey: 'categories',
                     searchFields: ['name'],
                     pagination: false,
+                    actionCallback: this.editCategory.bind(this),
                     viewOptions: {
                         table: {
                             openChildId: this.sandbox.sulu.getUserSetting(constants.lastClickedCategorySettingsKey),
-                            fullWidth: true,
                             selectItem: {
                                 type: 'checkbox',
                                 inFirstCell: true
                             },
                             icons: [
-                                {
-                                    column: 'name',
-                                    icon: 'pencil',
-                                    callback: this.editCategory.bind(this)
-                                },
                                 {
                                     column: 'name',
                                     icon: 'plus-circle',